PSTRESS-113 - Implement a script that converts pstress logs into SQL file
The script will be helpful in case we find a crash with pstress in single-thread mode. It will filter out SQLs
from the pstress thread logs and store it in a file called reduced.sql.

If a server is running, passing the socket information to the script will execute the SQL file sequentially against
a running server.
Script name: pstress_log_to_sql_converter.sh
Output File: reduced.sql
Usage:
./pstress_log_to_sql_converter.sh default.node.tld_step_1_thread-0.sql