PSTRESS-151 FIXES THE PARTITION INFO IN PSTRESS METADATA
https://jira.percona.com/browse/PSTRESS-151

The support for using MAXVALUE to create RANGE partition was
added recently. But, during the generation of CREATE TABLE
statement, the vector used to store partition information
wasn't being updated with the last(MAXVALUE) partition.

This issue is fixed and asserts are added to catch similar
issues if they pop up in the future.