Fix abort error when run with driver script when foreign keys are enabled
https://perconadev.atlassian.net/browse/PSTRESS-152

Problem
=======
The parent table pointer of the foreign key table object became invalid
after the first trial and was not updated properly for the subsequent
trials causing SIGSEGV errors.

Solution
========
The parent table name is also stored in the dll file for each step and
the parent table pointer for tables with foreign keys is set properly
when metadata is loaded from the dll files.